live: move iso/pxe customize infrastructure to a separate module
